+++ src/types/actor.d.ts
@@ -197,3 +197,5 @@ export interface DivineData {
 
 
 
+
+