+++ src/utils/calculations.ts
@@ -182,3 +182,5 @@ export function validateSkillValue(skillValue: number, masteryRank: number): num
 
 
 
+
+